Module remerge::schema::desc

Structs

Field
A single field in a record.

FieldIndex
Newtype wrapper around usize, referring into the fields vec in a RecordSchema

RecordSchema
The unserialized representation of the schema, parsed from a RawSchema (in json.rs). If you change this, you may have to change that as well.

Enums

ChangePreference

CompositeInfo

FieldKind

FieldType

IfOutOfBounds

TimestampDefault

TimestampSemantic

Constants

REMERGE_FEATURES_UNDERSTOOD
The set of features understood by this client.