Module remerge::schema::json

This module is concerned primarially with schema parsing (from RawSchema, e.g. the schema represented as JSON), and validation. It's a little bit hairy, and for the definitive documentation, you should refer to the docs/design/remerge/schema-format.md docs.

Structs

JustFormatVersion

RawFieldCommon

RawSchema
The serialized representation of the schema.

SchemaParser

TypeRestriction

Enums

ParsedMerge

RawFieldType

RawSpecialTime

RawTimeDefault
This (and RawSpecialTime) are basically the same as TimestampDefault, just done in such a way to make serde deserialize things the way we want for us.

RawTimestampSemantic

Constants

FORMAT_VERSION

Traits

BoundedNum

FieldErrorHelper
Helper trait to make marking results / errors with which field were were parsing more convenient.

Functions

compatible_version_req

is_valid_field_ident

parse_from_string

parse_version

parse_version_req

valid_origin_only_url