Module remerge::storage::records

This module provides wrappers around JsonValue that allow for better documentation and type safety for the format a method (usually in db.rs) is expected to take/return.

XXX The names "Local" vs "Native" here (and around) is confusing, but beats everything passing around serde_json::Values directly, and it matches the terms used in the RFC. I'm open to name suggestions, though.

Modules

private

sql_impls

Structs

LocalFormat
Record format for records in the current local schema. This is the format which we insert into the database, and it should always be newer or equal to the native format.

NativeFormat
A record in the native format understood by the local application using remerge. Data that comes from the FFI, and that is returned over the FFI should be in this format.

Record
A wrapper around serde_json::Value which indicates what format the record is in. Note that converting between formats cannot be done without schema information, so this is a paper-thin wrapper.

Traits

RecordFormat
Used to distinguish different categories of records.

Type Definitions

LocalRecord
A [Record] in [LocalFormat].

NativeRecord
A [Record] in [NativeFormat].