Module remerge::untyped_map

this module implements the untyped_map data type. This works as follows:

The "native" representation of an "untyped_map" value is just the underlying map data, e.g. a JsonObject.

To convert a native map to any other format of map, the tombstone list must be provided (of course, it's initially empty), creating an UntypedMap

To convert a local map to a native map, the tombstone gets stored in the database (if applicable) and then discarded. (Note: this happens in storage/records.rs, with the other local -> native conversion code)

See the RFC for the merge algorithm for these.

Structs

UntypedMap

Enums

CollisionHandling

OnCollision

Type Definitions

MapData